RSVP bug fix and Added Registry

+++ Data/Guest.cs
@@ -6,8 +6,8 @@ namespace Wedding.Data
     public class Guest
     {
         public Guid UserId { get; set; }
-        public int NumberAdults { get; set; }
-        public int NumberChildren { get; set; }
+        public byte NumberAdults { get; set; }
+        public byte NumberChildren { get; set; }
         public bool? IsGoing { get; set; }
         public string GuestName { get; set; }
         public string Email { get; set; }
