+++ Models/RsvpViewModel.cs
@@ -7,7 +7,7 @@ namespace Wedding.Models
         public string Name{get;set;}
         [Required]
         [Range(1, 20, ErrorMessage = "You must respond for at least one person.")]
-        public int NumberAttending{get;set;}
+        public byte NumberAttending{get;set;}
         [Required]
         [DataType(DataType.EmailAddress)]
         public string Email{get;set;}
